Added handling of errors and api response

+++ Valeting.API/Valeting.Services/BookingService.cs
@@ -1,7 +1,8 @@
 ﻿using Valeting.Business.Booking;
+using Valeting.Common.Exceptions;
 using Valeting.Services.Interfaces;
-using Valeting.Repositories.Interfaces;
 using Valeting.Business.VehicleSize;
+using Valeting.Repositories.Interfaces;
 
 namespace Valeting.Service
 {
@@ -18,6 +19,9 @@ namespace Valeting.Service
         {
             ValidateGeneralInput(bookingDTO);
 
+            if (bookingDTO.BookingDate < DateTime.Now)
+                throw new BusinessObjectException("Booking date cannot be in the past");
+
             bookingDTO.Id = Guid.NewGuid();
 
             await _bookingRepository.CreateAsync(bookingDTO);
@@ -30,12 +34,26 @@ namespace Valeting.Service
 
             ValidateBookingId(bookingDTO.Id);
 
+            if (!bookingDTO.Approved.HasValue)
+                throw new InputException("Approved is null or empty");
+
+            if (bookingDTO.BookingDate < DateTime.Now)
+                throw new BusinessObjectException("Booking date cannot be in the past");
+
+            BookingDTO bookingDTO_Check = await _bookingRepository.FindByIdAsync(bookingDTO.Id);
+            if (bookingDTO_Check == null)
+                throw new NotFoundException("Not found booking");
+
             await _bookingRepository.UpdateAsync(bookingDTO);
         }
 
         public async Task DeleteAsync(Guid id)
         {
             ValidateBookingId(id);
+
+            BookingDTO bookingDTO_Check = await _bookingRepository.FindByIdAsync(id);
+            if (bookingDTO_Check == null)
+                throw new NotFoundException("Not found booking");
 
             await _bookingRepository.DeleteAsync(id);
         }
@@ -44,9 +62,9 @@ namespace Valeting.Service
         {
             ValidateBookingId(id);
 
-            var bookingDTO = await _bookingRepository.FindByIDAsync(id);
+            var bookingDTO = await _bookingRepository.FindByIdAsync(id);
             if (bookingDTO == null)
-                throw new Exception("Booking not found");
+                throw new NotFoundException("Booking not found");
 
             return bookingDTO;
         }
@@ -54,7 +72,7 @@ namespace Valeting.Service
         public async Task<BookingListDTO> ListAllAsync(BookingFilterDTO bookingFilterDTO)
         {
             if (bookingFilterDTO.PageNumber == 0)
-                throw new Exception("pageNumber é 0");
+                throw new InputException("pageNumber é 0");
 
             return await _bookingRepository.ListAsync(bookingFilterDTO);
         }
@@ -62,31 +80,33 @@ namespace Valeting.Service
         private void ValidateGeneralInput(BookingDTO bookingDTO)
         {
             if (bookingDTO == null)
-                throw new Exception("bookingDTO is null");
+                throw new BusinessObjectException("bookingDTO is null");
 
             if (bookingDTO.Flexibility == null || bookingDTO.Flexibility.Id.Equals(Guid.Empty))
             {
                 string errorMsg = bookingDTO.Flexibility == null ? "Flexibility is null" : "FlexibilityId is empty";
-                throw new Exception(errorMsg);
+                throw new InputException(errorMsg);
             }
 
             if (bookingDTO.VehicleSize == null || bookingDTO.VehicleSize.Id.Equals(Guid.Empty))
             {
                 string errorMsg = bookingDTO.Flexibility == null ? "VehicleSize is null" : "VehicleSizeId is empty";
-                throw new Exception(errorMsg);
+                throw new InputException(errorMsg);
             }
 
-            if (string.IsNullOrEmpty(bookingDTO.Name) || string.IsNullOrEmpty(bookingDTO.Email))
+            if (string.IsNullOrEmpty(bookingDTO.Name) || string.IsNullOrEmpty(bookingDTO.Email) || !bookingDTO.ContactNumber.HasValue || bookingDTO.BookingDate.Equals(DateTime.MinValue))
             {
-                string errorMsg = string.IsNullOrEmpty(bookingDTO.Name) ? "Name is null or empty" : "Email is null or empty";
-                throw new Exception(errorMsg);
+                string errorMsg = string.IsNullOrEmpty(bookingDTO.Name) ? "Name is null or empty" :
+                                string.IsNullOrEmpty(bookingDTO.Email) ? "Email is null or empty" :
+                                !bookingDTO.ContactNumber.HasValue ? "Contact Number is null or empty" : "Booking date is null or empty";
+                throw new InputException(errorMsg);
             }
         }
 
         private void ValidateBookingId(Guid id)
         {
             if (id.Equals(Guid.Empty))
-                throw new Exception("BookingId is empty");
+                throw new InputException("BookingId is empty");
         }
     }
 }
